Add condition to not run user setup on GitHub actions

+++ setup.sh
@@ -1,14 +1,16 @@
 #!/bin/bash
-echo -n "Please choose a name for your project: "
-read PROJECT_NAME
-find ./ -type f -exec sed -i "s/exampleproject/$PROJECT_NAME/g" {} \;
-mv exampleproject $PROJECT_NAME
-echo "# $PROJECT_NAME" > README.md
-head -1 setup.sh > /tmp/setup.sh
-tail -3 setup.sh >> /tmp/setup.sh
-rm LICENSE setup.sh
-cp /tmp/setup.sh .
-chmod +x setup.sh
+if [ -z "$GITHUB_ACTIONS" ]; then
+    echo -n "Please choose a name for your project: "
+    read PROJECT_NAME
+    find ./ -type f -exec sed -i "s/exampleproject/$PROJECT_NAME/g" {} \;
+    mv exampleproject $PROJECT_NAME
+    echo "# $PROJECT_NAME" > README.md
+    head -1 setup.sh > /tmp/setup.sh
+    tail -3 setup.sh >> /tmp/setup.sh
+    rm LICENSE setup.sh
+    cp /tmp/setup.sh .
+    chmod +x setup.sh
+fi
 python3 -m venv venv
 source venv/bin/activate
 pip install -r requirements.txt
